Split encoding core into udigest-encoding
Introduce udigest-encoding as the dependency-free encoding crate and make udigest a facade that re-exports encoding APIs and provides digest-based hash helpers.
