+++ udigest-encoding/CHANGELOG.md
@@ -1,0 +1,2 @@
+## v0.10.0
+- Initial split from udigest as dependency-free encoding core
