+++ udigest-encoding/src/_macros.rs
