+++ udigest-encoding/src/inline_struct.rs
